DB/DWBA: Refactored Born approximation/subtraction.
First Born approximation routines have been separated from "hex-dwba"
into a common file. Born subtraction in "hex-db" has been reintroduced.
It will be used for correction of dipole transitions whenever "hex-db"
is compiled with support for GiNaC. In such case, evaluation of dipole
transitions will take noticeably longer time, as both the total Born
amplitude and (particularly) the partial Born amplitudes are evaluated
on the fly. This is universal and works for both the legacy and current
"hex-db" databases.