ECS: BOINC support.
Hex-ecs is made compatible to run within the BOINC infrastructure.
This particularly includes the ability to checkpoint the calculation
on every iteration and recover from these checkpoints. In the stand-alone
mode one can trigger checkpoint (and exit) by invoking SIGINT, or
the checkpoints can be made automatically with the option
'--checkpoints'. Recovery from the checkpoint is automatic if '--continue'
option is used for the next start. This will also delete the checkpoint.
Further, hex-ecs now supports the option '-nthreads' issued by BOINC
for multithreaded applications of class "mt".