Drive by dipole-coupled custom solution
The right-hand side of the driven Schrodinger equation can be now built
using an earlier solution and the electric potential in the velocity
gauge (V = polarization . momentum). This is useful to calculate one-
and multi-photon ionization using the Dalgarno-Lewis approach discussed
in Phys. Rev. A 103, 043108 (2021).

To use such a right-hand side, use the `--rhs-dipV <input>` command line
option with path to the input file used for calculation of the driving
solution. Of course, the B-spline grids should be compatible. It is
allowed that they are formally not identical. Then the B-spline elements
of the driving solution will be simply copied over to the new grid.
This is useful when the driving state is a bound state calculated on the
real part of the grid only (or even on a subset of the grid). However,
positive-energy (scattering) states should definitely use the same
radial grids.

It is assumed that the polarization is directed along z-axis.

The input file should then contain only one initial state (ideally 1s).
The initial state is ignored in the calculation, but this avoids
repeated calculations with the same source term. For each total energy
in the input file, the driving solution for the same energy *index* will
be loaded.

If there are fewer driving solutions, they will be read cyclically. This
is useful when the driving solution is the bound state of the system,
which is just one. For the driven solutions it will be used repeatedly.
It is expected that the total energy is specified as "nan" in the input
file for the bound state.